--- main/model/ScatsData-Bundoora/bilstm_loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.0014893421903252602,132481.171875,0.0010454547591507435,247467.375
-0.0010503126541152596,108379.125,0.000896054320037365,170749.671875
-0.0010032334830611944,107311.2890625,0.0008419679943472147,47605.51171875
