--- main/model/ScatsData-Bundoora/cnnlstm_loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.00304982578381896,242776.0,0.002162287477403879,167121.265625
-0.0021520229056477547,161221.5,0.0019223518902435899,110651.828125
-0.0020125403534621,153440.96875,0.0018202783539891243,239176.671875
