--- main/model/ScatsData-Bundoora/gru_loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.002755442401394248,338308.59375,0.0010370333911851048,138434.296875
-0.0011073725763708353,167517.609375,0.0009558843448758125,108040.65625
-0.0010364808840677142,138992.03125,0.0009313964983448386,94096.0546875
