--- main/model/ScatsData-Bundoora/lstm_loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.003662631381303072,366871.59375,0.0010523685486987233,164685.59375
-0.0011282430496066809,160813.828125,0.0008804082754068077,116880.9609375
-0.0010419663740321994,135153.65625,0.0009899032302200794,187260.34375
