--- main/model/ScatsData-Bundoora/saes_loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.003988249227404594,305169.03125,0.0013722013682126999,253285.84375
-0.0012927062343806028,159400.234375,0.0010447244858369231,106706.859375
-0.0012032765662297606,138080.109375,0.0010037533938884735,107618.015625
