--- main/model/TrafficCountLocation/bilstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.0025829761289060116,27728.890625,0.001881855190731585,10611.5
-0.001648641424253583,11167.451171875,0.0017235741252079606,4352.35693359375
-0.001490647904574871,10183.2197265625,0.0015662062214687467,7373.955078125
