--- main/model/TrafficCountLocation/cnnlstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.0020644327159971,18248.73828125,0.0018839059630408883,12822.9619140625
-0.0016108941053971648,8539.853515625,0.001398964668624103,9300.50390625
-0.0015255151083692908,7319.99609375,0.001754957134835422,11564.03515625
