--- main/model/TrafficCountLocation/gru loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.003840931924059987,30898.041015625,0.002392283407971263,9441.203125
-0.0019067818066105247,9095.16796875,0.0016210294561460614,10365.3876953125
-0.0017140859272331,7427.18310546875,0.0019647073931992054,28580.626953125
