--- main/model/TrafficCountLocation/lstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.0037521927151829004,29731.544921875,0.0020455464255064726,5526.6650390625
-0.0018123447662219405,7369.541015625,0.0018154801800847054,4275.94384765625
-0.0016884597716853023,5369.13916015625,0.0017019326332956553,5185.7353515625
