--- main/model/TrafficCountLocation/saes loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.0029425998218357563,16669.615234375,0.0017142245778813958,13420.6962890625
-0.001735231257043779,6955.759765625,0.0017149500781670213,2406.337890625
-0.001627580844797194,3935.9775390625,0.0016638411907479167,2475.021728515625
