--- main/model/default/bilstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.015376667492091656,58428.5625,0.01575559750199318,31.58978843688965
-0.008400717750191689,32339.845703125,0.005623538978397846,39.164424896240234
-0.006556713487952948,32785.0859375,0.008829804137349129,27.062179565429688
