--- main/model/default/cnnlstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.03831348195672035,76522.40625,0.009525912813842297,70.48546600341797
-0.009493732824921608,58778.3515625,0.008273317478597164,34.94856262207031
-0.00897595752030611,39820.59765625,0.00806821696460247,36.04043960571289
