--- main/model/default/gru loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.021150724962353706,166086.4375,0.00700064841657877,61.2641487121582
-0.006983999162912369,88680.34375,0.006293893326073885,36.12963104248047
-0.0058988407254219055,54872.375,0.005038455128669739,45.40259552001953
