--- main/model/default/lstm loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.02511344477534294,273082.375,0.008866130374372005,45.933406829833984
-0.009775885380804539,54215.6640625,0.008345512673258781,44.75965881347656
-0.008609088137745857,38490.5390625,0.00819576159119606,29.242576599121094
