--- main/model/default/saes loss.csv
@@ -1,4 +1,0 @@
-loss,mape,val_loss,val_mape
-0.06092750281095505,202893.609375,0.032016851007938385,307.7995300292969
-0.030160363763570786,127599.78125,0.027960149571299553,65.37208557128906
-0.01972355507314205,67217.4375,0.012850913219153881,122.75415802001953
