+++ main/model/trained/lstm_loss.csv
@@ -1,0 +1,4 @@
+loss,mape,val_loss,val_mape
+0.003606132697314024,372552.53125,0.0009947018697857857,127731.8359375
+0.0011302935890853405,165891.046875,0.000993300462141633,115587.1015625
+0.0010408165398985147,135509.171875,0.0009137931047007442,90997.28125
