+++ main/trainSCATS.py
@@ -26,9 +26,9 @@ def train_model(model, X_train, y_train, name, config):
         epochs=config["epochs"],
         validation_split=0.05)
 
-    model.save(f'model/ScatsData-Bundoora/{name}.h5')
+    model.save(f'model/trained/{name}.h5')
     df = pd.DataFrame.from_dict(hist.history)
-    df.to_csv(f'model/ScatsData-Bundoora/{name}_loss.csv', encoding='utf-8', index=False)
+    df.to_csv(f'model/trained/{name}_loss.csv', encoding='utf-8', index=False)
 
 def train_seas(models, X_train, y_train, name, config):
     """train
